Backend/server/requirements.txt
Django==2.2.9
httplib2==0.9.1
idna==2.0
psycopg2-binary==2.8.3
pytz==2019.2
django-bootstrap4==0.0.8
django-cors-headers==3.2.1
django-rest-auth==0.9.5
djangorestframework==3.11.0
